requirements.txt
asgiref==3.3.1
boto==2.49.0
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-debug-toolbar==3.2
django-material-admin==1.7.16
django-storages==1.11.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
jmespath==0.10.0
openpyxl==3.0.7
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycountry==20.7.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
